refactor: find repository scene
