+++ app/src/main/java/com/example/user_repository_infos/AppModule.kt
@@ -4,4 +4,5 @@ import org.koin.dsl.module
 
 val appModule = module {
     viewModel { FindRepositoryViewModel() }
+    viewModel { FindUserInfoViewModel() }
 }
